Interactable Controller Script Finished!
Have finally finished the script that handles the outlining of the object!

It works like this;
-Player Enters Object Range
 Grab outline width and set it to a smooth lerp to a predetermined value (currently called 'TargetOutlineWidth'

-Player Stays Within Object Range
Object outline will oscillate from TargetOutlineWidth +/- Amplitude, with Frequency controlling speed

-Player Leaves Object Range
Outline will quickly lerp back to 0

Ive spent a fair bit of time allowing it to be interrupted by breaking the Lerp In/Out coroutines if the player enters and exits quickly.

The player also has 2 hitboxes on them, so I've added a Hitbox Counter, so only the outermost hitbox will affect the outline.

Currently the outliner cannot handle more than 1 interactable object, and will only outline the object that overlapped with the hitbox first.