Eye Following
after a really long time stressing over how to get this done, its finally complete.

initially I was restricting both eyes individually until I realised that was a really stupid idea and that I should parent both pupils to a central driver, and limit that.

But now its ok! Using the InteractableObjects script to grab the current object and using MoveTowards to aim the eyes.

I did have to use an empty GameObject to direct the eyes back to the original position? which feels a little amateur, but it works.

Id like for the eyelids to be driven using a script, with the max height being set to either half way up the eye OR the top of the pupil, so if the player needs to look up the eyelids will easily account for that. I could also use that to drive certain emotions like 'intrigued', where only one eyelid & eyebrow slightly raises.

but I think ill work on designing the conductor before then. Im happy with how things are going